+++ Source/Scene/Globe.js
@@ -148,6 +148,12 @@ import TileSelectionResult from './TileSelectionResult.js';
         this.enableLighting = false;
 
         /**
+         * Enable lighting effects on atmosphere and fog. This only takes effect
+         * when <code>enableLighting</code> is true.
+         */
+        this.enableAtmosphereLighting = true;
+
+        /**
          * Enable the ground atmosphere, which is drawn over the globe when viewed from a distance between <code>lightingFadeInDistance</code> and <code>lightingFadeOutDistance</code>.
          *
          * @demo {@link https://sandcastle.cesium.com/index.html?src=Ground%20Atmosphere.html|Ground atmosphere demo in Sandcastle}
@@ -177,7 +183,8 @@ import TileSelectionResult from './TileSelectionResult.js';
 
         /**
          * The distance where the darkness of night from the ground atmosphere fades out to a lit ground atmosphere.
-         * This only takes effect when <code>showGroundAtmosphere</code> and <code>enableLighting</code> are <code>true</code>.
+         * This only takes effect when <code>showGroundAtmosphere</code>, <code>enableLighting</code>, and
+         * <code>enableAtmosphereLighting</code> are <code>true</code>.
          *
          * @type {Number}
          * @default 10000000.0
@@ -186,7 +193,8 @@ import TileSelectionResult from './TileSelectionResult.js';
 
         /**
          * The distance where the darkness of night from the ground atmosphere fades in to an unlit ground atmosphere.
-         * This only takes effect when <code>showGroundAtmosphere</code> and <code>enableLighting</code> are <code>true</code>.
+         * This only takes effect when <code>showGroundAtmosphere</code>, <code>enableLighting</code>, and
+         * <code>enableAtmosphereLighting</code> are <code>true</code>.
          *
          * @type {Number}
          * @default 50000000.0
@@ -741,6 +749,7 @@ import TileSelectionResult from './TileSelectionResult.js';
             tileProvider.hasWaterMask = hasWaterMask;
             tileProvider.oceanNormalMap = this._oceanNormalMap;
             tileProvider.enableLighting = this.enableLighting;
+            tileProvider.enableAtmosphereLighting = this.enableAtmosphereLighting;
             tileProvider.showGroundAtmosphere = this.showGroundAtmosphere;
             tileProvider.shadows = this.shadows;
             tileProvider.hueShift = this.atmosphereHueShift;