+++ Source/Scene/GlobeSurfaceTileProvider.js
@@ -84,6 +84,7 @@ import TileSelectionResult from './TileSelectionResult.js';
         this.oceanNormalMap = undefined;
         this.zoomedOutOceanSpecularIntensity = 0.5;
         this.enableLighting = false;
+        this.enableAtmosphereLighting = false;
         this.showGroundAtmosphere = false;
         this.shadows = ShadowMode.RECEIVE_ONLY;
 
@@ -1494,6 +1495,7 @@ import TileSelectionResult from './TileSelectionResult.js';
         showReflectiveOcean : undefined,
         showOceanWaves : undefined,
         enableLighting : undefined,
+        enableAtmosphereLighting : undefined,
         showGroundAtmosphere : undefined,
         perFragmentGroundAtmosphere : undefined,
         hasVertexNormals : undefined,
@@ -1657,6 +1659,7 @@ import TileSelectionResult from './TileSelectionResult.js';
         surfaceShaderSetOptions.showReflectiveOcean = showReflectiveOcean;
         surfaceShaderSetOptions.showOceanWaves = showOceanWaves;
         surfaceShaderSetOptions.enableLighting = tileProvider.enableLighting;
+        surfaceShaderSetOptions.enableAtmosphereLighting = tileProvider.enableAtmosphereLighting;
         surfaceShaderSetOptions.showGroundAtmosphere = showGroundAtmosphere;
         surfaceShaderSetOptions.perFragmentGroundAtmosphere = perFragmentGroundAtmosphere;
         surfaceShaderSetOptions.hasVertexNormals = hasVertexNormals;