+++ Source/Scene/Scene.js
@@ -2866,7 +2866,7 @@ import View from './View.js';
             environmentState.moonCommand = undefined;
         } else {
             if (defined(skyAtmosphere) && defined(globe)) {
-                skyAtmosphere.setDynamicAtmosphereColor(globe.enableLighting);
+                skyAtmosphere.setDynamicAtmosphereColor(globe.enableLighting && globe.enableAtmosphereLighting);
                 environmentState.isReadyForAtmosphere = environmentState.isReadyForAtmosphere || globe._surface._tilesToRender.length > 0;
             }
             environmentState.skyAtmosphereCommand = defined(skyAtmosphere) ? skyAtmosphere.update(frameState) : undefined;